Added Zoom, Fixed Error Message
Made the following changes:
Added zoom to set command
Fixed error message for single-argument plot command

On single-window display issue:
Qt4Agg backend does not include figure manager, according to yt (see base_plot_types.PlotMPL._set_canvas), although the backend_qt5.FigureManagerQT class would appear to do the trick -
perhaps using Qt5Agg would work. What therefore displays when slc.show() is called is the PlotMPL's canvas, which is a C/C++ object. The canvas does not appear to allow access to its window, so the commented out
add_toolbar method cannot be used to modify the window, and the toolbar controls may not work regardless.

The commented out code in plot_cmd is as close as I have gotten - it displays the plot in a single window with the toolbar when slc.show is commented out, but the controls cannot interact
properly with the canvas, including save (says a C/C++ object of type FigureCanvasQTAgg was deleted and does not save anything).

First version should be complete, however - it is just reliant on the command line for interacting with the plot.